Improve output directory creation
Previously, if an output directory was specified on the command line, only the current directory from which the script was run would be checked for the directory to be created to see if it already existed. This method of directory creation could easily raise exceptions.

Now, the script always attempts to create the directory. If an exception is raised, errno is checked to see if a file-exists error triggered the exception. If so, then the script continues, and the already existing directory is used.